+++ ink/thulac/include/punctuation.h
@@ -30,7 +30,7 @@ public:
     void adjust(SegmentedSentence& sentence){
         if(!p_dat)return;
         std::vector<Word> tmpVec;
-        for(int i = 0 ; i < sentence.size(); i ++){
+        for(int i = 0 ; i < (int)sentence.size(); i ++){
             Word tmp = sentence[i];
             if(p_dat->get_info(tmp) >= 0) continue;
 
@@ -38,7 +38,7 @@ public:
 
             tmpVec.clear();
             int j;
-            for(j = i + 1; j < sentence.size(); j ++){
+            for(j = i + 1; j < (int)sentence.size(); j ++){
                 tmp += sentence[j];
                 if(p_dat->get_info(tmp) >= 0){
                     break;
@@ -70,7 +70,7 @@ public:
         if(!p_dat)return;
         std::vector<Word> tmpVec;
 	bool findMulti = false;
-        for(int i = 0 ; i < sentence.size(); i ++){
+        for(int i = 0 ; i < (int)sentence.size(); i ++){
             Word tmp = sentence[i].word;
             if(p_dat->get_info(tmp) >= 0) continue;
 
@@ -78,7 +78,7 @@ public:
 
             tmpVec.clear();
             int j;
-            for(j = i + 1; j < sentence.size(); j ++){
+            for(j = i + 1; j < (int)sentence.size(); j ++){
                 tmp += sentence[j].word;
                 if(p_dat->get_info(tmp) >= 0){
                     break;
